rest_api_user/requirements.txt
python-dotenv
requests
Flask-SQLAlchemy
flask
psycopg2
